price as int64 fix

+++ backend/internal/Domain/models/Product.go
@@ -3,7 +3,7 @@ package models
 type Product struct {
 	ID       int
 	Name     string
-	Price    int
+	Price    int64
 	Category string
 	ImageURL string
 }
